Fix the way arrays are implemented
Before arrays sometimes were defined directly as a stack variable pointing to the first element and sometimes as a pointer pointing to the first element. Now it's always the former. It's always a stack variable pointing directly to the first element unless the user creates a pointer themselves (by using malloc for example)